version 2

+++ Project1/EthernetPacket.h
@@ -20,7 +20,7 @@ private:
 	
 
 public:
-	EthernetPacket(std::vector<uint8_t> srcMacAdress_, std::vector<uint8_t> destMacAdress_, std::vector<uint8_t> payLoad_, uint8_t etherType_);
+	EthernetPacket(std::vector<uint8_t> srcMacAdress_, std::vector<uint8_t> destMacAdress_, std::vector<uint8_t> payLoad_, uint16_t etherType_);
 	EthernetPacket(std::vector<uint8_t> srcMacAdress_, std::vector<uint8_t> destMacAdress_,
 		uint8_t payloadValue, int numBytes, uint16_t etherType_);  //this constructor will generate the same data
 	std::vector<uint8_t> GetPacket();
